Check for existence of schema before dropping

+++ lib/go_cart/target.rb
@@ -1,6 +1,5 @@
 require 'active_record'
 require 'activerecord-import'
-#require 'logger'
 
 module GoCart
 class Target
@@ -33,8 +32,12 @@ class Target
 	def drop_schema(dbconfig, dialect, schema)
 		ActiveRecord::Base.establish_connection(dbconfig)
 		begin
-			# TODO: Move to dialect, this is Postgresql specific
-			ActiveRecord::Base.connection.execute "DROP SCHEMA #{schema} CASCADE;"
+			unless @schema.nil?
+				# TODO: Move to dialect, this is Postgresql specific
+				if ActiveRecord::Base.connection.schema_exists?(@schema)
+					ActiveRecord::Base.connection.execute "DROP SCHEMA #{schema} CASCADE;"
+				end
+			end
 		ensure
 			ActiveRecord::Base.connection.disconnect!
 		end
@@ -42,7 +45,6 @@ class Target
 
 	def open_database_connection(dbconfig)
 		ActiveRecord::Base.establish_connection(dbconfig)
-		#ActiveRecord::Base.logger = Logger.new(STDERR)
 		unless @schema.nil?
 			# TODO: Move to dialect, this is Postgresql specific
 			unless ActiveRecord::Base.connection.schema_exists?(@schema)
